requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
requests==2.31.0
apscheduler==3.10.4
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
bencodepy==0.9.5
